+++ Handler/PLC/Logging/SqlClasses/LogValuesSql.cs
@@ -4,7 +4,7 @@ namespace OptimaValue
 {
     public class LogValuesSql
     {
-        public decimal tag_id { get; set; }
+        public int tag_id { get; set; }
         public DateTime logTime { get; set; }
         public string value { get; set; }
         public float numericValue { get; set; }
