+++ Handler/PLC/Logging/SqlClasses/TagDefinitions.cs
@@ -6,7 +6,7 @@ namespace OptimaValue
 {
     public class TagDefinitions : TagStatistics
     {
-        public decimal id { get; set; }
+        public int id { get; set; }
         public bool active { get; set; }
         public string name { get; set; }
         public LogType logType { get; set; }
@@ -31,15 +31,15 @@ namespace OptimaValue
         #endregion
 
         // Events
-        public decimal eventId { get; set; } // The id of the trigger tag
+        public int eventId { get; set; } // The id of the trigger tag
         public bool IsBooleanTrigger { get; set; }
         public BooleanTrigger boolTrigger { get; set; }
         public AnalogTrigger analogTrigger { get; set; }
 
         public float analogValue { get; set; } = 0f;
 
-        private List<decimal> subscribedTags = new List<decimal>();
-        public List<decimal> SubscribedTags
+        private List<int> subscribedTags = new List<int>();
+        public List<int> SubscribedTags
         {
             get
             {
